fix riscv macros

+++ vfw_core/include/vfw_arch/riscv/vfw_riscv.h
@@ -35,35 +35,35 @@ extern void wait_mcycle64(uint64_t cnt);
 })
 
 #define read_csr(x) ({ \
-    unsigned int r = 0; \
+    uintptr_t r = 0; \
     __asm__ volatile ("csrrs %0, %1, x0" : "=r" (r) : "i"((x))); \
     r; \
 })
 
 #define write_csr(x, v) ({ \
-    unsigned int r = (v); \
+    uintptr_t r = (v); \
     __asm__ volatile ("csrrw x0, %1, %0" : "+r" (r) : "i"((x))); \
 })
 
 #define set_csr(x, v) { \
-    unsigned int r = (v); \
+    uintptr_t r = (v); \
     __asm__ volatile ("csrrs x0, %1, %0" : "+r" (r) : "i"((x))); \
 }
 
 #define clr_csr(x, v) { \
-    unsigned int r = (v); \
+    uintptr_t r = (v); \
     __asm__ volatile ("csrrc x0, %1, %0" : "+r" (r) : "i"((x))); \
 }
 
 #define get_sp() ({ \
-    register unsigned int sp; \
+    register uintptr_t sp; \
     __asm__ volatile ("mv %0, sp" : "=r" (sp)); \
     sp; \
 })
 
 #define check_stack() ({ \
-    unsigned int start = stack_start(); \
-    unsigned int end = start - stack_size(); \
+    uintptr_t start = stack_start(); \
+    uintptr_t end = start - stack_size(); \
     assert((get_sp() <= start) && (get_sp() > end)); \
 })
